backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "david",email = "david.guerth@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.75.0,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "boto3 (>=1.38.13,<2.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


